requirements.txt
# Core
streamlit==1.28.0
pandas==2.1.3
numpy==1.26.0

# Extraction texte
pdfminer.six==20221105
python-docx==0.8.11
PyPDF2==3.0.1

# NLP
spacy==3.7.2
fr-core-news-md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.7.0/fr_core_news_md-3.7.0-py3-none-any.whl
nltk==3.8.1
transformers==4.34.0
sentence-transformers==2.2.2

# Vectorisation et similarité
scikit-learn==1.3.2
scipy==1.11.3

# Visualisation
matplotlib==3.8.0
wordcloud==1.9.2
plotly==5.18.0

# Gestion fichiers
openpyxl==3.1.2
python-docx==0.8.11

protobuf==3.20.3